# api/openapi.yaml
openapi: 3.1.0
info:
  title: NovaForge API
  version: 0.1.0
  description: >-
    The NovaForge control plane. The web GUI is deferred, so this contract is
    the complete description of the platform for every client: the nf CLI,
    external agents, and the future GUI.
servers:
  - url: /
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    Unauthorized:
      description: No usable credentials were presented
    Forbidden:
      description: The caller's organization scope does not permit this
    NotFound:
      description: No such resource in the caller's organization
paths:
...
  /api/v1/orgs/{org}/repos/{repo}:
    delete:
      operationId: deleteRepo
      summary: Delete a repository
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      operationId: getRepo
      summary: One repository
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/orgs/{org}/repos/{repo}/ci/artifacts:
    get:
      operationId: listLatestArtifacts
      summary: The newest run's artifacts
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/ci/artifacts/{id}:
    get:
      operationId: downloadArtifact
      summary: Download one artifact's content
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/ci/jobs/{id}/artifacts:
    get:
      operationId: listJobArtifacts
      summary: One job's artifacts
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/ci/jobs/{id}/logs:
    get:
      operationId: getJobLogs
      summary: One job's log
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/ci/logs:
    get:
      operationId: getLatestJobLogs
      summary: The newest run's first job log
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/ci/runs:
    get:
      operationId: listCIRuns
      summary: CI runs for a repository
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: triggerCIRun
      summary: Run the repository's workflow at a ref
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/orgs/{org}/repos/{repo}/ci/runs/{id}:
    get:
      operationId: getCIRun
      summary: One CI run and its jobs
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/orgs/{org}/repos/{repo}/runs:
    get:
      operationId: listRuns
      summary: Engineering Runs
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createRun
      summary: Open an Engineering Run
      parameters:
        - name: org
          in: path
          required: true
          schema:
            type: string
        - name: repo
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Success
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...